+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/application/CircuitBreakerService.java
@@ -1,16 +1,53 @@
 package com.mageddo.dnsproxyserver.solver.remote.application;
 
+import com.mageddo.commons.circuitbreaker.CircuitCheckException;
+import com.mageddo.commons.circuitbreaker.CircuitIsOpenException;
 import com.mageddo.dnsproxyserver.solver.remote.CircuitStatus;
 import com.mageddo.dnsproxyserver.solver.remote.Result;
+import com.mageddo.dnsproxyserver.solver.remote.application.failsafe.CircuitBreakerFactory;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.ClassUtils;
 
+import javax.inject.Inject;
+import javax.inject.Singleton;
 import java.net.InetSocketAddress;
 import java.util.function.Supplier;
 
-public interface CircuitBreakerService {
 
-  // fixme #533 esse padrão de strategy não é mais necessário aqui, foi movido para CircuitBreakerDelegate
-  //  onde tem mais chances de reduzir duplicação
-  Result safeHandle(final InetSocketAddress resolverAddress, Supplier<Result> sup);
+@Slf4j
+@Singleton
+@RequiredArgsConstructor(onConstructor = @__({@Inject}))
+public class CircuitBreakerService {
 
-  CircuitStatus findCircuitStatus(InetSocketAddress resolverAddress);
+  private final CircuitBreakerFactory circuitBreakerFactory;
+
+  private String status;
+
+  public Result safeHandle(InetSocketAddress resolverAddress, Supplier<Result> sup) {
+    try {
+      return this.handle(resolverAddress, sup);
+    } catch (CircuitCheckException | CircuitIsOpenException e) {
+      final var clazz = ClassUtils.getSimpleName(e);
+      log.debug("status=circuitEvent, server={}, type={}", resolverAddress, clazz);
+      this.status = String.format("%s for %s", clazz, resolverAddress);
+      return Result.empty();
+    }
+  }
+
+  private Result handle(InetSocketAddress resolverAddress, Supplier<Result> sup) {
+    return this.circuitBreakerFactory.check(resolverAddress, sup);
+  }
+
+  public String getStatus() {
+    return this.status;
+  }
+
+  public void resetCircuitBreakerFactory() {
+    this.circuitBreakerFactory.reset();
+  }
+
+  public CircuitStatus findCircuitStatus(InetSocketAddress resolverAddress) {
+    return this.circuitBreakerFactory.findStatus(resolverAddress);
+  }
 }
