+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/configurator/SolverRemoteModule.java
@@ -1,7 +1,5 @@
 package com.mageddo.dnsproxyserver.solver.remote.configurator;
 
-import com.mageddo.dnsproxyserver.solver.remote.application.CircuitBreakerService;
-import com.mageddo.dnsproxyserver.solver.remote.application.failsafe.CircuitBreakerFailSafeService;
 import com.mageddo.dnsproxyserver.solver.remote.dataprovider.SolverConsistencyGuaranteeDAO;
 import com.mageddo.dnsproxyserver.solver.remote.dataprovider.SolverConsistencyGuaranteeDAOImpl;
 import dagger.Binds;
@@ -11,9 +9,6 @@ import javax.inject.Singleton;
 
 @Module
 public interface SolverRemoteModule {
-  @Binds
-  @Singleton
-  CircuitBreakerService circuitBreakerService(CircuitBreakerFailSafeService impl);
 
   @Binds
   @Singleton
